requirements.txt
gradio==3.37.0
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
h5py